feat: created the page folder,  moved home & cart
